BookCapture/requirements.txt
pyautogui>=0.9.54
mss>=9.0.1
Pillow>=10.0.0
numpy>=1.24.0
pyinstaller>=6.0.0
